requirements.txt
Django==5.2.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
Pillow==10.4.0
boto3==1.35.0
google-auth==2.35.0
requests==2.32.5
django-cors-headers==4.3.1